Move collect_data_for_plugin() and related constants to plugins.py
