+++ .github/scripts/plugins.py
@@ -4,8 +4,10 @@ import re
 import os
 import sys
 
-from utils import get_template
+from utils import get_template, get_plugin_manifest
 
+MOBILE_COMPATIBLE = "[[Mobile-compatible plugins|Yes]]"
+DESKTOP_ONLY = "[[Desktop-only plugins|No]]"
 
 CORE_PLUGINS = [
     {
@@ -160,6 +162,32 @@ def get_core_plugins():
 
     with open(file_path, "w") as md_file:
         md_file.write(new_contents)
+
+
+def collect_data_for_plugin(plugin, file_groups):
+    """
+    Take raw plugin data from a community plugin, and add information to it,
+    typically from its manifest file.
+
+    :param plugin: A dict with data about the plugin, to be updated by this function
+    :param file_groups: Place to store error message if the plugin is invalid
+    :return: Whether the plugin is valid, and is OK to be added to the Hub
+    """
+    repo = plugin.get("repo")
+    branch = plugin.get("branch", "master")
+    manifest = get_plugin_manifest(repo, branch)
+
+    plugin_is_valid = validate_plugin(plugin, manifest, repo, file_groups)
+
+    user = repo.split("/")[0]
+    if manifest.get("isDesktopOnly"):
+        mobile = DESKTOP_ONLY
+    else:
+        mobile = MOBILE_COMPATIBLE
+
+    plugin.update(mobile=mobile, user=user, **manifest)
+
+    return plugin_is_valid
 
 
 def validate_plugin(plugin, manifest, repo, file_groups):
